requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
decorator==4.4.2
distro==1.9.0
dnspython==2.7.0
fastapi==0.115.5
fastapi-login==1.10.2
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
motor==3.6.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai==1.54.4
pillow==11.0.0
proglog==0.1.10
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pymongo==4.9.2
python-dotenv==1.0.1
python-multipart==0.0.17
regex==2024.11.6
requests==2.32.3
setuptools==75.5.0
sniffio==1.3.1
starlette==0.41.2
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
